refactor: shift tests to admin-forms.logic.routes.spec
